fix: actually use leading ones
We claimed to switch to leading_ones in
c3177af289bfd8862e43f507b81317b60c9038c0, but that was actually not
true. The commit removed the inverting, but kept leading_zeros, which
means that it broke the library.

To prevent further breakage like this going unnoticed, we'll use CI for
publishing from now on, which means a release will only get onto
crates.io after the CI pipeline succeeded